Split classes from GrainInterfaceMap.cs
